+++ packages/tenoxui/src/types/index.ts
@@ -18,3 +18,31 @@ export interface Config extends CoreConfig {
   prefixLoaderOptions: MoxieConfig
   reservedVariantChars: string[]
 }
+
+export type ShorthandItem = {
+  className: string
+  cssRules: string | string[] | null
+  value: string | null
+  isImportant: boolean
+  variants: null | { name: string; data: string | null }
+  raw: null | (string | undefined)[]
+}
+
+export type AliasItem = {
+  className: string
+  rules: { cssRules: string | string[] | null; value: string | null; isImportant: boolean }[]
+  isImportant: boolean
+  prefix: null | { name: string; data: string | null }
+  variants:
+    | null
+    | {
+        className: string
+        rules: {
+          cssRules: string | string[] | null
+          value: string | null
+          isImportant: boolean
+        }[]
+        variant: string
+      }[]
+  raw: null | (string | undefined)[]
+}